Amelioration of Serialize and fix error
Change std::vector<>& to std::vector
change Define PRECISION to serialize_precision
Fix stod to stoi
Change Factory vector to int in parameter

+++ enki/robots/e-puck/EPuck.cpp
@@ -142,7 +142,7 @@ namespace Enki
 
 	void EPuck::serialize(std::ostream& oss, const bool first) const
 	{
-		oss << static_cast<int>(Factory::TypeObject::EPUCK) << TYPE_SEPARATOR << getId() << TYPE_SEPARATOR;
+		oss << Factory::TypeObject::EPUCK << TYPE_SEPARATOR << getId() << TYPE_SEPARATOR;
 		Robot::serializeRobot(oss);
 		oss << OBJECT_SEPARATOR;
 	}
