+++ enki/robots/marxbot/Marxbot.cpp
@@ -75,7 +75,7 @@ namespace Enki
 	}
 
 	void Marxbot::serialize(std::ostream& oss, const bool first) const{
-		oss << static_cast<int>(Factory::TypeObject::MARXBOT) << TYPE_SEPARATOR<< getId() << TYPE_SEPARATOR;
+		oss << Factory::TypeObject::MARXBOT << TYPE_SEPARATOR<< getId() << TYPE_SEPARATOR;
 		Robot::serializeRobot(oss);
 		oss << OBJECT_SEPARATOR;
 	}
